Dao interfaces removed. Methods implemented into Dao classes. Useless code removed.
