+++ src/main/java/com/mmdmanager/servlets/AdminDashboard.java
@@ -1,7 +1,6 @@
 package com.mmdmanager.servlets;
 
 import com.mmdmanager.dao.AdminDAO;
-import com.mmdmanager.dao.AdminDAOImpl;
 import com.mmdmanager.others.Admin;
 
 import javax.servlet.ServletException;
@@ -11,7 +10,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.sql.*;
 
 @WebServlet("/AdminDashboard")
 public class AdminDashboard extends HttpServlet {
@@ -23,7 +21,7 @@ public class AdminDashboard extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         doGet(request, response);
 
-        adminDAO = new AdminDAOImpl();
+        adminDAO = new AdminDAO();
 
         admin_id = request.getParameter("userID");
         acc_password = request.getParameter("userPassword");
